Test ERC20 reclaiming & fix nil dereference in saveFailedTokenDeposit()
